backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.23
psycopg2-binary>=2.9.9
alembic>=1.12.1
python-multipart>=0.0.6
python-dotenv>=1.0.0
openpyxl>=3.1.2
qrcode[pil]>=7.4.2
pillow>=10.1.0
opencv-python-headless>=4.8.1.78
pyzbar>=0.1.9
websockets>=12.0
cryptography>=41.0.7
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
pydantic>=2.12.5
pydantic-settings>=2.12.0